requirements.txt
kafka-python-ng
pymongo==4.6.1
requests==2.31.0